work around conversion problem by changing the long array to an int

+++ src/main/java/dllWrapper/ResultWrapper.java
@@ -1,6 +1,6 @@
 package dllWrapper;
 
 public class ResultWrapper {
-    public long[] intensities;
-    public long[] indicies;
+    public int[] intensities;
+    public int[] indicies;
 }
